requirements.txt
appdirs==1.4.4
argcomplete==3.1.2
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
MarkupSafe==2.1.3
Pillow==10.0.1
pyodbc==5.0.0
pypng==0.20220715.0
pyserial==3.5
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==6.0.1
qrcode==7.4.2
six==1.16.0
typing_extensions==4.8.0
Werkzeug==3.0.0
